Implemented ev3 download and refactored error checking
- ev3 download didn’t work because maxBytes was 4 bytes instead of 4
hexadecimal digits (2 bytes).
- communication errors are now handled in main.c

+++ src/packets.h
@@ -118,7 +118,7 @@ typedef struct
 {
 	EV3_COMMAND_FIELDS
 
-	u32 maxBytes;
+	u16 maxBytes;
 	char fileName[];
 } BEGIN_UPLOAD;
 
@@ -138,7 +138,7 @@ typedef struct
 	EV3_COMMAND_FIELDS
 
 	u8 fileHandle;
-	u32 maxBytes;
+	u16 maxBytes;
 } CONTINUE_UPLOAD;
 	
 //! doesn't work
